+++ NEWS.md
@@ -7,6 +7,7 @@
 - network_data now handles deleted accounts (#730)
 - retryonratelimit works again on paginated endpoints (#732)
 - Fixed a bug in `auth_sitrep` due to duplicate tokens
+- Provided `rbind` and subsetting methods (`[`) for rtweet new classes
 
 # rtweet 1.0.2